+++ src/verse-neigh.cpp
@@ -171,12 +171,12 @@ void Train() {
                  << step / (float)(total_steps + 1) * 100 << "%";
         last_ncount = ncount;
       }
-      int n1 = irand(nv);
-      int n2 = sample_neighbor(n1);
+      size_t n1 = irand(nv);
+      size_t n2 = sample_neighbor(n1);
       if(n2 == -1) continue;
       update(&w0[n1 * n_hidden], &w0[n2 * n_hidden], 1, nce_bias);
       for (int i = 0; i < n_samples; i++) {
-        int neg = irand(nv);
+        size_t neg = irand(nv);
         update(&w0[n1 * n_hidden], &w0[neg * n_hidden], 0, nce_bias_neg);
       }
       ncount++;
@@ -245,7 +245,7 @@ int main(int argc, char **argv) {
     return 0;
   }
   w0 = static_cast<float *>(aligned_malloc(nv * n_hidden * sizeof(float), DEFAULT_ALIGN));
-  for (int i = 0; i < nv * n_hidden; i++)
+  for (size_t i = 0; i < nv * n_hidden; i++)
     w0[i] = drand() - 0.5;
   degrees = (int *)malloc(nv * sizeof(int));
   for (int i = 0; i < nv; i++)
@@ -261,7 +261,7 @@ int main(int argc, char **argv) {
        << chrono::duration_cast<std::chrono::duration<float>>(end - begin)
               .count()
        << " s to run" << endl;
-  for (int i = 0; i < nv * n_hidden; i++)
+  for (size_t i = 0; i < nv * n_hidden; i++)
     if (w0[0] != w0[0]) {
       cout << endl << "NaN! Not saving the result.." << endl;
       return 1;
